% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eefAnalyticPerm.R
\name{catcht}
\alias{catcht}
\title{Multisite trial data.}
\format{A data frame with 210 rows and 4 variables}
\description{
A multisite education trial data containing 54 schools.
}
\details{
\itemize{
  \item Posttest. posttest scores
  \item Prettest. prettests scores
  \item Intervention. indicator for intervention groups
coded as 1 for the intervention and 0 for the control group.
  \item School. numeric school identifier
}
}

