% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefAnalytics_14_09_2020.R
\name{crtData}
\alias{crtData}
\title{Cluster Randomised Trial Data.}
\format{
A data frame with 265 rows and 5 variables
}
\description{
A cluster randomised trial dataset containing 22 schools. The data contains a random sample of test data of pupils and not actual trial data.
}
\details{
\itemize{
  \item Posttest: posttest scores
  \item Prettest: prettest scores
  \item Intervention: the indicator for intervention groups in a two arm trial,
coded as 1 for intervention group and 0 for control group. The intervention variable should always be numeric with the control group as the first value.
  \item Intervention2: a simulated indicator for intervention groups in a three arm trial. The intervention variable should always be numeric with the control group as the first value.
  \item School: numeric school identifier
}
}
