% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eefAnalyticPerm.R
\name{crtFREQ}
\alias{crtFREQ}
\title{Analysis of Cluster Randomised Trials using MLM.}
\usage{
crtFREQ(formula, random, intervention, nPerm = NULL, nBoot = NULL, data)
}
\arguments{
\item{formula}{specifies in the model using form posttest ~ pretests+Intervention +....}

\item{random}{a string variable specifying the "clustering"  as contained in the data.}

\item{intervention}{specifies the intervention variable in the formula.}

\item{nPerm}{specifies number of perumation under the null hypothesis.  Default is NULL.}

\item{nBoot}{specifies number of bootstrap samples for non-parametric bootstrapping. Default is NULL.}

\item{data}{specifies data frame containing the data to be analysed.}
}
\value{
S3 object; a list consisting of
\itemize{
\item \code{Beta}. The parameter estimates from linear model with the associatie confidence intervals.
\item \code{ES}. A matric of  Hedges effect size based on within variance and total variance. It is base is based on standard error (SE) if \code{nBoot=NULL}. Otherwise, it is based on non-parametric bootstrap confidence intervals.
\item \code{covParm}. Vector of variance decomposition into between-variance (Schools) and within-variance (Pupils).
\item \code{SchEffects}. Estimated random effects for each school.
\item \code{Perm}. A matrix containing the distribution of effect size under the null hypothesis. It is only produced if \code{nPerm} is specified. The first columns contains effect size based on within variance and the second column contains effect size based on total variance.
\item \code{Bootstrap}. A matrix containing the bootstrap effect sizes. It is only prduced is \code{nBoot} is specified. The first columns contains effect size based on within variance and the second column contains effect size based on total variance.
}
}
\description{
\code{crtFREQ} performs analysis of cluster randomised trials.
}
\examples{
data(iwq)

###############################################
## MLM analysis of cluster randomised trials ##
###############################################

output1 <- crtFREQ(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",data=iwq)

ES1 <- output1$ES$Intervention1

ICC <- output1$covParm[1]/sum(output1$covParm)
ICC

################################################################
## MLM analysis of cluster randomised trials with permutation ##
## under the null hypothesis                                  ##
################################################################

date()
output2 <- crtFREQ(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",nPerm=1000,data=iwq)
ES2 <- output2$ES$Intervention1
ES2

ICC2 <- output2$covParm[1]/sum(output2$covParm)
ICC2

date()

perm <- output2$Perm


#### Within variance

obsg <- output2$ES$Intervention1[1,1]

tp <- ifelse(mean(perm[,1] > obsg[1])==0,"<0.001",mean(perm[,1] > obsg[1]) )

hist(perm[,1], breaks=40, col="white", border="blueviolet", 
		xlab="Distribution Under Null Hypothesis", main=paste("P(X|NULL)= ",tp,sep=""),
		xlim=range(c(perm[,1],obsg),na.rm=TRUE));
abline(v=obsg[1],lwd=2,col=4)


#### Total variance

obsg <- output2$ES$Intervention1[2,1]

tp <- ifelse(mean(perm[,2] > obsg[1])==0,"<0.001",mean(perm[,2] > obsg[1]) )

hist(perm[,1], breaks=40, col="white", border="blueviolet",
		xlab="Distribution Under Null Hypothesis", 
		main=paste("P(X|NULL)= ",tp,sep=""), xlim=range(c(perm[,1],obsg),
				na.rm=TRUE))
abline(v=obsg[1],lwd=2,col=4)

###############################################
## MLM analysis of cluster randomised trials ##	 
## with bootstrap confidence intervals       ##
###############################################

output3 <- crtFREQ(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",nBoot=1000,data=iwq)
ES3 <- output3$ES
ES3


ICC3 <- output3$covParm[1]/sum(output3$covParm)
ICC3

#########################################################################
## Bayesian MLM analysis of cluster randomised trials with permutation ##
## under the null hypothesis                                           ##
#########################################################################


output4 <- crtBayes(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",nSim=10000,data=iwq)
ES4 <- output4$ES
ES4



}

