% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecate.R
\name{eefAnalytics-defunct}
\alias{eefAnalytics-defunct}
\alias{mlmbayes}
\alias{caceMSTBoot}
\alias{caceCRTBoot}
\alias{caceSRTBoot}
\title{Defunct functions in eefAnalytics}
\usage{
mlmbayes(...)

caceMSTBoot(...)

caceCRTBoot(...)

caceSRTBoot(...)
}
\description{
These functions are marked as defunct and have been removed from eefAnalytics.

These functions are marked as defunct and have been removed from eefAnalytics.

These functions are marked as defunct and have been removed from eefAnalytics.

These functions are marked as defunct and have been removed from eefAnalytics.
}
\keyword{internal}
