% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefAnalyticPerm.R
\name{iwq}
\alias{iwq}
\title{Cluster randomised trial data.}
\format{A data frame with 265 rows and 5 variables}
\description{
A cluster randomised trial data containing 22 schools.
}
\details{
\itemize{
  \item Posttest. posttest scores
  \item Prettest. prettests scores
  \item Intervention. indicator for intervention groups 
coded as 1 for the intervention group and 0 for the control group.
  \item Compliance. percentage of sessions attended by pupils
  \item School. numeric school identifier
}
}

