% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eefAnalyticPerm.R
\name{mees.plot}
\alias{mees.plot}
\title{Minimum expected effect size plot.}
\usage{
mees.plot(output, x.cord = 0.8, y.cord = max(esParm2) - 0.05)
}
\arguments{
\item{output}{\code{crtBayes} outout object}

\item{x.cord}{position of legend on x-axis.}

\item{y.cord}{position of legend on y-axis.}
}
\value{
A plot of minimum expected effect size.
}
\description{
\code{crtFREQ} performs analysis of cluster randomised trials.
}
\examples{
data(iwq)
#########################################################################
## Bayesian MLM analysis of cluster randomised trials with permutation ##
## under the null hypothesis                                           ##
#########################################################################


output4 <- crtBayes(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",nSim=10000,data=iwq)

mees.plot(output=output4,x.cord=0.0,y.cord=0.2)

}

