% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefAnalytics_03_2017.r
\name{mlmBayes}
\alias{mlmBayes}
\title{Bayesia multilevel analysis of randomised educatuon trials using vague priors.}
\usage{
mlmBayes(formula, random, intervention, nSim, data)
}
\arguments{
\item{formula}{specifies the model to be analysed. It is of the form y ~ x1+x2+....}

\item{random}{a string variable specifying the "cluster variable"  as contained in the data.}

\item{intervention}{a string variable specifying the intervention variable in the formula.}

\item{nSim}{specifies the number of MCMC iterations. A minimum of 10,000 is recommended.}

\item{data}{specifies a data frame to be analysed.}
}
\value{
S3 object; a list consisting of
\itemize{
\item \code{Beta}. Parameter estimates from the linear model with associated confidence intervals.
\item \code{ES}. A matrix of  Hedges' g based on within and total variances. It is based on standard errors (SE) if \code{nBoot=NULL}. Otherwise, it is non-parametric bootstrapped confidence intervals.
\item \code{covParm}. Vector of variance decomposition into between-variance (Schools) and within-variance (Pupils).
\item \code{SchEffects}. Estimated random intercepts for individual schools.
\item \code{ProbES}. A maxtrix containing the probability of observing ES greater than a pre-specified value. First column is for within-variance, second for between-variance and the third for total-variance.
}
}
\description{
\code{mlmBayes} performs analysis of randomised eduation trials using multilevel model under Bayesian framework 
assuming vague priors.
}
\examples{
if(interactive()){
  
data(crtData)

########################################################
## Bayesian analysis of cluster randomised trials     ##
########################################################

output <- mlmBayes(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",nSim=10000,data=crtData)

### Fixed effects
beta <- output$Beta
beta

### Effect size
ES1 <- output$ES
ES1

## Covariance matrix
covParm <- output$covParm
covParm

### plot random effects for schools

plotObject(analyticObject=output)

### plot posterior probability of an effect size to be bigger than a pre-specified threshold

plotObject(analyticObject=output,group=1)

}


}

