% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefAnalytics_14_09_2020.R
\name{mstData}
\alias{mstData}
\title{Multisite Trial Data.}
\format{
A data frame with 210 rows and 5 variables
}
\description{
A multisite trial dataset containing 54 schools. This data contains a random sample of test data of pupils and not actual trial data.
}
\details{
\itemize{
  \item Posttest: posttest scores
  \item Prettest: prettest scores
  \item Intervention: the indicator for the intervention groups in a two arm trial,
coded as 1 for intervention group and 0 for control group. The intervention variable should always be numeric with the control group as the first value.
  \item Intervention2: a simulated indicator for intervention groups in a three arm trial. The intervention variable should always be numeric with the control group as the first value.
  \item School: numeric school identifier
}
}
