% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefAnalyticPerm.R
\name{mstFREQ}
\alias{mstFREQ}
\title{Analysis of Multisite Randomised Trials using MMLM.}
\usage{
mstFREQ(formula, random, intervention, nPerm = NULL, data, nBoot = NULL)
}
\arguments{
\item{formula}{specifies the model to be analysed. It is of the form form y ~ x1+x2 +..., 
where y is the outcome variable and X's are the predictors.}

\item{random}{a string variable specifying the "clustering" variable as contained in the data.
This must be put between  quotes. For example, "school".}

\item{intervention}{specifies the name of the intervention variable as appeared in formula.
This must be put between quotes.  For example "intervention" or "treatment" or "group"...}

\item{nPerm}{number of permutations required to generate permutation p-value. Default  is NULL.}

\item{data}{the data frame to be analysed.}

\item{nBoot}{number of bootstrap required to generate bootstrap confidence interval. Default is NULL.}
}
\value{
S3 object; a list consisting of
\itemize{
\item \code{Beta}. Estimates and confidence intervals for the predictors specified in the model. 
It will be the slope for a continuous predictor and the mean difference for a dummy variable or a categorical predictor.
\item \code{ES}. Hedges' effect size for the intervention effect. If nBoot is not specified, 
the confidence intervals are 95% CIs based on standard error. If nBoot is specified, 
they are non-parametric bootstrapped confidence intervals.
\item \code{covParm}. A vector of variance decomposition into between-variance (Schools), within-variance (Pupils) and total variance.
It also contains the intra-cluster correlation (ICC).
\item \code{SchEffects}. Individual school effects at baseline.
\item \code{Perm}. A matrix of the distribution of ES under the null hypothesis. 
The two columns in the matrix represents ES based on within variance and total variance. 
Produced only if nPerm is specified.
}
}
\description{
\code{mstFREQ} performs multilevel model for analysing randomised multisite trials with schools and school by intervention interactions specified as random effects
}
\examples{
data(catcht)

########################################################
## MLM analysis of multisite trials + 1.96SE ##
########################################################

output1 <- mstFREQ(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",data=catcht)


### Fixed effects
beta <- output1$Beta
beta

### Effect size
ES1 <- output1$ES$Intervention1
ES1

## Covariance matrix
covParm <- output1$covParm
covParm

### random effects for schools

randOut <- output1$"SchEffects"
randOut <- randOut[order(randOut$Estimate),]
barplot(randOut$Estimate,ylab="Deviations from Overall Average",
		names.arg=randOut$Schools,las=2)

###############################################
## MLM analysis of multisite trials          ##	 
## with bootstrap confidence intervals       ##
###############################################

output2 <- mstFREQ(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",nBoot=1000,data=catcht)


### Effect size

ES2 <- output2$ES
ES2


#######################################################################
## MLM analysis of mutltisite trials with permutation p-value##
#######################################################################

output3 <- mstFREQ(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",nPerm=1000,data=catcht)


#### Distribution under the null

perm <- output3$Perm
str(perm )

#### Permutation P-value using total variance


obsg <- output3$ES$Intervention1[2,1]
obsg


p_value <- ifelse(mean(perm$"Intervention1Total" > obsg)==0,"<0.001",
		mean(perm$"Intervention1Total" > obsg) )
p_value

hist(perm$"Intervention1Total", breaks=40, col="white", 
		border="blueviolet",
		xlab="Distribution Under Null Hypothesis", 
		main=paste("P(X|NULL)= ",p_value,sep=""), 
		xlim=range(c(perm$"Intervention1Total",obsg),
		na.rm=TRUE))
abline(v=obsg,lwd=2,col=4)





}

