% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_Summary.R
\name{print.eefAnalytics}
\alias{print.eefAnalytics}
\title{Print for a fitted model represented by an \code{eefAnalytics} object.}
\usage{
\method{print}{eefAnalytics}(x, ...)
}
\arguments{
\item{x}{Object of class \code{eefAnalytics}}

\item{...}{Additional arguments of \code{\link[base]{print}}}
}
\value{
Print conditional and unconditional effect sizes.
}
\description{
Print for a fitted model represented by an \code{eefAnalytics} object.
}
