% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefAnalyticPerm_modified_05_02_2016.R
\name{caceMSTBoot}
\alias{caceMSTBoot}
\title{CACE Analysis of Multisite Randomised Trials.}
\usage{
caceMSTBoot(formula, random, intervention, compliance, nBoot, data)
}
\arguments{
\item{formula}{model specification of the form posttest ~ pretests+Intervention+.... the model to be analysed. 
It is of the form y~x1+x2+...,  where y is the outcome variable and X's are the predictors.}

\item{random}{a string variable specifying the "clustering" variable as contained in the data.
This must be put between  quotes. For example, "school".}

\item{intervention}{the name of the intervention variable as appeared in formula.
This must be put in quotes.  For example "intervention" or "treatment" or "group".}

\item{compliance}{percentages of sessions attended by pupils.}

\item{nBoot}{number of bootstrap required to generate bootstrap confidence interval. This must be specified.}

\item{data}{data frame containing the data to be analysed.}
}
\value{
S3 object; a list consisting of
\itemize{
\item \code{CACE}. Estimated CACE effect size based on percentages of sessions attended by pupils. 
The percentage data is converted into the following grids (0, 10, 20, 30, 40, 50, 60, 70, 80, 90) 
and CACE effect size is calculated for each grid.
\item \code{Compliers}. A summary table of the percentage of pupils in the intervention and control groups that 
attended more than a pre-specified percentage of sessions. The values for the control group should be zeros if 
there is no dilution in which a pupil or school in the control group receives intervention.
}
}
\description{
\code{caceMSTBoot} performs CACE analysis of multisite randomised trials.
}
\examples{
data(catcht)

########################################################
## MLM analysis of multisite trials + 1.96SE ##
########################################################

output1 <- mstFREQ(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",data=catcht)


### Fixed effects
beta <- output1$Beta
beta

### Effect size
ES1 <- output1$ES$Intervention1
ES1

## Covariance matrix
covParm <- output1$covParm
covParm

### random effects for schools

randOut <- output1$"SchEffects"
randOut <- randOut[order(randOut$Estimate),]
barplot(randOut$Estimate,ylab="Deviations from Overall Average",
		names.arg=randOut$Schools,las=2)

###############################################
## MLM analysis of multisite trials          ##	 
## with bootstrap confidence intervals       ##
###############################################

output2 <- mstFREQ(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",nBoot=1000,data=catcht)


### Effect size

ES2 <- output2$ES
ES2


#######################################################################
## MLM analysis of mutltisite trials with permutation p-value##
#######################################################################

output3 <- mstFREQ(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",nPerm=1000,data=catcht)


#### Distribution under the null

perm <- output3$Perm
str(perm )

#### Permutation P-value using total variance


obsg <- output3$ES$Intervention1[2,1]
obsg


p_value <- ifelse(mean(perm$"Intervention1Total" > obsg)==0,"<0.001",
		mean(perm$"Intervention1Total" > obsg) )
p_value

hist(perm$"Intervention1Total", breaks=40, col="white", 
		border="blueviolet",
		xlab="Distribution Under Null Hypothesis", 
		main=paste("P(X|NULL)= ",p_value,sep=""), 
		xlim=range(c(perm$"Intervention1Total",obsg),
		na.rm=TRUE))
abline(v=obsg,lwd=2,col=4)





}

