% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{ComparePlot}
\alias{ComparePlot}
\title{A plot function to compare diferent eefAnalytics S3 objects from the eefAnalytics package.}
\usage{
ComparePlot(eefAnalyticsList, group = NULL, modelNames = NULL)
}
\arguments{
\item{eefAnalyticsList}{A list of eefAnalytics S3 objects from eefAnalytics package.}

\item{group}{a value indicating which intervention to plot.
This must be not be greater than the number of intervention excluing the control group.
For a two arm trial, the maximum value is 1 and a maximum value of 2 for three arm trial.}

\item{modelNames}{a string factor, which must be specified, containing the names of model to compare.}
}
\value{
Returns a forest plot to compare the different models
}
\description{
\code{ComparePlot} plots a forest plot comparing the different eefAnalytics results.
}
\details{
MAYBE A BIT MORE DETAILS ABOUT WHAT INFORMATION CAN BE FOUND ON THE FOREST PLOT?
}
\examples{
if(interactive()){

#### read data
data(mstData)
data(crtData)


###############
##### SRT #####
###############

##### Two arm study 
outputSRT <- srtFREQ(Posttest~ Intervention + Prettest,
                     intervention = "Intervention", data = mstData)
# plot(outputSRT)


outputSRTBoot <- srtFREQ(Posttest~ Intervention + Prettest,
                         intervention = "Intervention",nBoot=1000, data = mstData)
# plot(outputSRTBoot)
plot(outputSRTBoot,group=1)


outputSRTPerm <- srtFREQ(Posttest~ Intervention + Prettest,
                         intervention = "Intervention",nPerm=1000, data = mstData)
# plot(outputSRTPerm)
plot(outputSRTPerm,group=1)


##### Three arm study 
outputSRT2 <- srtFREQ(Posttest~ Intervention2 + Prettest,
                      intervention = "Intervention2", data = mstData)
# plot(outputSRT2)


outputSRTBoot2 <- srtFREQ(Posttest~ Intervention2 + Prettest,
                          intervention = "Intervention2",nBoot=1000, data = mstData)
# plot(outputSRTBoot2)
plot(outputSRTBoot2,group=1)
plot(outputSRTBoot2,group=2)


outputSRTPerm2 <- srtFREQ(Posttest~ Intervention2 + Prettest,
                          intervention = "Intervention2",nPerm=1000, data = mstData)
# plot(outputSRTPerm2)
plot(outputSRTPerm2,group=1)
plot(outputSRTPerm2,group=2)

###############
##### MST #####
###############


#### Two arm
outputMST <- mstFREQ(Posttest~ Intervention + Prettest, 
                     random = "School", intervention = "Intervention", data = mstData)
plot(outputMST)


outputMSTBoot <- mstFREQ(Posttest~ Intervention + Prettest, 
                         random = "School", intervention = "Intervention", 
                         nBoot = 1000, data = mstData)
plot(outputMSTBoot)
plot(outputMSTBoot,group=1)


outputMSTPerm <- mstFREQ(Posttest~ Intervention + Prettest, 
                         random = "School", intervention = "Intervention",
                         nPerm = 1000, data = mstData)
plot(outputMSTPerm)
plot(outputMSTPerm,group=1)


outputMSTbayes <- mlmBayes(Posttest~ Intervention + Prettest, 
                           random = "School", intervention = "Intervention",
                           nSim = 10000, data = mstData)
plot(outputMSTbayes)
plot(outputMSTbayes,group=1)

## comparing different results

ComparePlot(eefAnalyticsList=list(outputSRT,outputSRTBoot,outputMST,outputMSTBoot,outputMSTbayes),
            modelNames =c("ols", "olsBoot","MLM","MLMBoot","MLMBayes"),group=1)
# ComparePlot(eefAnalyticsList=list(outputSRT,outputSRTBoot,outputMST,outputMSTBoot,outputMSTbayes),
#             modelNames =c("ols", "olsBoot","MLM","MLMBoot","MLMBayes"))

#####

#### Three arm
outputMST2 <- mstFREQ(Posttest~ Intervention2 + Prettest, random = "School", 
                      intervention = "Intervention2", data = mstData)
plot(outputMST2)


outputMSTBoot2 <- mstFREQ(Posttest~ Intervention2 + Prettest, random = "School", 
                          intervention = "Intervention2", nBoot = 1000, data = mstData)
plot(outputMSTBoot2,group=1)
plot(outputMSTBoot2,group=2)


outputMSTPerm2 <- mstFREQ(Posttest~ Intervention2 + Prettest, random = "School", 
                          intervention = "Intervention2", nPerm = 1000, data = mstData)
plot(outputMSTPerm2,group=1)
plot(outputMSTPerm2,group=2)


outputMSTbayes2 <- mlmBayes(Posttest~ Intervention2 + Prettest, random = "School", 
                            intervention = "Intervention2", nSim = 10000, data = mstData)
plot(outputMSTbayes2)
plot(outputMSTbayes2,group=1)
plot(outputMSTbayes2,group=2)


ComparePlot(eefAnalyticsList=
              list(outputSRT2,outputSRTBoot2,outputMST2,outputMSTBoot2,outputMSTbayes2),
            modelNames =c("ols", "olsBoot","MLM","MLMBoot","MLMBayes"),group=1)
ComparePlot(eefAnalyticsList=
              list(outputSRT2,outputSRTBoot2,outputMST2,outputMSTBoot2,outputMSTbayes2),
            modelNames =c("ols", "olsBoot","MLM","MLMBoot","MLMBayes"),group=2)



###############
##### CRT #####
###############

#### Two arm
outputCRT <- crtFREQ(Posttest~ Intervention + Prettest, random = "School", 
                     intervention = "Intervention", data = crtData)
plot(outputCRT)


outputCRTBoot <- crtFREQ(Posttest~ Intervention + Prettest, random = "School", 
                         intervention = "Intervention", nBoot = 1000, data = crtData)
plot(outputCRTBoot,group=1)


outputCRTPerm <- crtFREQ(Posttest~ Intervention + Prettest, random = "School", 
                         intervention = "Intervention", nPerm = 1000, data = crtData)
plot(outputCRTPerm,group=1)


outputCRTbayes <- mlmBayes(Posttest~ Intervention + Prettest, random = "School", 
                           intervention = "Intervention", nSim = 10000, data = crtData)
plot(outputCRTbayes,group=1)


ComparePlot(eefAnalyticsList=list(outputCRT,outputCRTBoot ,outputCRTbayes),
            modelNames=c("Std Error","BootStrap","Bayesian"),group=1)



#### Three arm
outputCRT2 <- crtFREQ(Posttest~ Intervention2 + Prettest, random = "School", 
                      intervention = "Intervention2", data = crtData)
plot(outputCRT2)


outputCRTBoot2 <- crtFREQ(Posttest~ Intervention2 + Prettest, random = "School", 
                          intervention = "Intervention2", nBoot = 1000, data = crtData)
plot(outputCRTBoot2,group=1)
plot(outputCRTBoot2,group=2)


outputCRTPerm2 <- crtFREQ(Posttest~ Intervention2 + Prettest, random = "School", 
                          intervention = "Intervention2", nPerm = 1000, data = crtData)
plot(outputCRTPerm2,group=1)
plot(outputCRTPerm2,group=2)

outputCRTbayes2 <- mlmBayes(Posttest~ Intervention2 + Prettest, random = "School", 
                            intervention = "Intervention2", nSim = 10000, data = crtData)
plot(outputCRTbayes2,group=1)
plot(outputCRTbayes2,group=2)


ComparePlot(eefAnalyticsList=list(outputCRT2,outputCRTBoot2 ,outputCRTbayes2),
            modelNames=c("Std Error","BootStrap","Bayesian"),group=1)
ComparePlot(eefAnalyticsList=list(outputCRT2,outputCRTBoot2 ,outputCRTbayes2),
            modelNames=c("Std Error","BootStrap","Bayesian"),group=2)


################
##### CACE #####
################

outputSRTCace <- caceSRTBoot(Posttest~ Intervention + Prettest,
                             intervention = "Intervention",compliance="Percentage_Attendance",
                             nBoot=1000, data = mstData) 
plot(outputSRTCace)
plot(outputSRTCace,group=2)


outputMSTCace <- caceMSTBoot(Posttest~ Intervention + Prettest,random="School",
                             intervention = "Intervention",compliance="Percentage_Attendance",
                             nBoot=1000, data = mstData) 
plot(outputMSTCace)


outputCRTCace <- caceCRTBoot(Posttest~ Intervention + Prettest,random="School",
                             intervention = "Intervention",compliance="Percentage_Attendance",
                             nBoot=1000, data = crtData) 
plot(outputCRTCace)


}
}

