% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefAnalyticPerm_modified_05_02_2016.R
\name{crtBayes}
\alias{crtBayes}
\title{Bayesia analysis of cluster randomised trials Using vague priors.}
\usage{
crtBayes(formula, random, intervention, nSim, data)
}
\arguments{
\item{formula}{specifies the model to be analysed. It is of the form form y ~ x1+x2 +..., 
where y is the outcome variable and X's are the predictors.}

\item{random}{a string variable specifying the "clustering" variable as contained in the data.
This must be put between  quotes. For example, "school".}

\item{intervention}{specifies the name of the intervention variable as appeared in formula.
This must be put between quotes.  For example "intervention" or "treatment" or "group"..}

\item{nSim}{number of MCMC simulations to generate samples from full conditional posterior distributions. 
A minimum of 10,000 is recommended.}

\item{data}{specifies data frame containing the data to be analysed.}
}
\value{
S3 \code{mcpi} object; a list consisting of
\itemize{
\item \code{Beta}. Estimates and confidence intervals for the predictors specified in the model. 
It will be a slope for a continuous predictor and a mean difference for a dummy variable or a categorical predictor.
\item \code{ES}. Effect size for the intervention effect.
\item \code{covParm}. A vector of variance decomposition into between-variance (Schools), within-variance (Pupils) and total variance.
It also contains the intra-cluster correlation (ICC).
\item \code{ProbES}. A maxtrix containing the probability of observing ES greater than a pre-specified value. First column is for within-variance, second column for between-variance and the third column for total-variance.
\item \code{SchEffects}. Individual school effects at baseline.
}
}
\description{
\code{crtBayes} performs analysis of cluster randomised trial using multilevel model within the Bayesian framework 
assuming vague priors.
}
\examples{
data(iwq)

########################################################
## Bayesian analysis of cluster randomised trials     ##
########################################################

output <- crtBayes(Posttest~ Intervention+Prettest,
		random="School",intervention="Intervention",
		nSim=10000,data=iwq)

### Fixed effects
beta <- output$Beta
beta

### Effect size
ES1 <- output$ES
ES1

## Covariance matrix
covParm <- output$covParm
covParm

### random effects for schools

randOut <- output$"SchEffects"
randOut <- randOut[order(randOut$Estimate),]
barplot(randOut$Estimate,ylab="Deviations from Overall Average",
		names.arg=randOut$Schools,las=2)



### Posterior probability given a fixed threshold
probES <- output$ProbES
str(probES )

			
plot(probES[,1] ,probES[,2],ylim=c(0,max(probES)),
		ylab="Probability",cex.lab=1,cex.axis=1,
		type="n", xlab=expression("Effect size" >= "x"),
		cex=1)
lines(probES[,1],probES[,2],col="chartreuse3",cex=1.5,
		lwd=1.5,lty=2)
lines(probES[,1],probES[,3],col="violetred",cex=1.5,
		lwd=1.5,lty=3)
lines(probES[,1],probES[,4],col="cornflowerblue",cex=1.5,
		lwd=1.5,lty=1)
points(probES[,1],probES[,2],col="chartreuse3",cex=1.5,
		lwd=1.5,pch=7)
points(probES[,1],probES[,3],col="violetred",cex=1.5,
		lwd=1.5,pch=1)
points(probES[,1],probES[,4],col="cornflowerblue",
		cex=1.5,lwd=1.5,pch=12)
legend(0,0.4,legend=c("Within ","Between ","Total "),
		lty=c(2,3,1),cex=1.5, pch=c(7,1,12),
		col=c("chartreuse3","violetred","cornflowerblue"),
		title="Variance Type")




}

