% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eefAnalyticPerm.R
\name{crtBayes}
\alias{crtBayes}
\title{Bayesia analysis of cluster randomised trials Using vague priors.}
\usage{
crtBayes(formula, random, intervention, nSim, data)
}
\arguments{
\item{formula}{specifies in the model using form posttest ~ pretests+Intervention+....}

\item{random}{a string variable specifying the "clustering"  as contained in the data.}

\item{intervention}{specifies the intervention variable in the formula.}

\item{nSim}{specifies number of MCMC iterations. A minimum of 10,000 is recommended.}

\item{data}{specifies data frame containing the data to be analysed.}
}
\value{
S3 \code{mcpi} object; a list consisting of
\itemize{
\item \code{Beta}. The parameter estimates from linear model with the associatie confidence intervals.
\item \code{ES}. A matric of  Hedges effect size based on within variance and total variance. It is base is based on standard error (SE) if \code{nBoot=NULL}. Otherwise, it is based on non-parametric bootstrap confidence intervals.
\item \code{covParm}. Vector of variance decomposition into between-variance (Schools) and within-variance (Pupils).
\item \code{SchEffects}. Estimated random effects for each school.
\item \code{ProbES}. A maxtrix containing the probability of observing ES greater than a pre-specified value. First column is for within-variance, second column for between-variance and the third column for total-variance.
}
}
\description{
\code{crtBayes} performs Bayesian analysis of clustered randomised trials using vague priors.
}
\examples{
data(iwq)

#########################################################################
## Bayesian MLM analysis of cluster randomised trials with permutation ##
## under the null hypothesis                                           ##
#########################################################################

date()
output4 <- crtBayes(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",nSim=10000,data=iwq)
ES4 <- output4$ES
ES4

date()


}

