% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefAnalytics_03_2017.r
\name{crtData}
\alias{crtData}
\title{Cluster randomised trial data.}
\format{A data frame with 265 rows and 5 variables}
\description{
A cluster randomised trial dataset containing 22 schools.
}
\details{
\itemize{
  \item Posttest. posttest scores
  \item Prettest. prettest scores
  \item Intervention. indicator for intervention groups 
  \item Intervention2. a simulated indicator for intervention groups three arm trial 
coded as 1 for intervention group and 0 for control group.
  \item Compliance. percentage of sessions attended by pupils
  \item School. numeric school identifier
}
}

