% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefAnalytics_03_2017.r
\name{crtFREQ}
\alias{crtFREQ}
\title{Analysis of Cluster Randomised Education Trials using Multilevel Model.}
\usage{
crtFREQ(formula, random, intervention, nPerm = NULL, nBoot = NULL, data)
}
\arguments{
\item{formula}{specifies the model to be analysed. It is of the form y ~ x1+x2+....}

\item{random}{a string variable specifying the "clustering variable"  as contained in the data.}

\item{intervention}{a string variable specifying the intervention variable in the formula.}

\item{nPerm}{specifies number of perumation under the null hypothesis.  Default is NULL.}

\item{nBoot}{specifies number of bootstrap samples for non-parametric bootstrap confidence interval. Default is NULL.}

\item{data}{specifies a data frame containing the data to be analysed.}
}
\value{
S3 object; a list consisting of
\itemize{
\item \code{Beta}. The parameter estimates from multilevel model with their associated confidence intervals.
\item \code{ES}. A matrix of  Hedges effect size based on within variance and total variance. It is base is based on standard error (SE) if \code{nBoot=NULL}. Otherwise, it is based on non-parametric bootstrap confidence intervals.
\item \code{covParm}. Vector of variance decomposition into between-variance (Schools) and within-variance (Pupils).
\item \code{SchEffects}. Estimated random intercepts for individual schools.
\item \code{Perm}. A matrix containing the distribution of effect sizes under the null hypothesis. It is only produced when \code{nPerm} is specified. The first column contains effect sizes based on within variance and the second contains effect sizes based on total variance.
\item \code{Bootstrap}. A matrix containing bootstrapped effect sizes. It is only prduced when \code{nBoot} is specified. The first column contains effect sizes based on within variance and the second one contains effect sizes based on total variance.
}
}
\description{
\code{crtFREQ} performs Analysis of cluster randomised education trial using multilevel model under the frequentist framework.
}
\examples{
if(interactive()){

data(crtData)

########################################################
## MLM analysis of cluster randomised trials + 1.96SE ##
########################################################

output1 <- crtFREQ(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",data=crtData)


### Fixed effects
beta <- output1$Beta
beta

### Effect size
ES1 <- output1$ES$Intervention1
ES1

## Covariance matrix
covParm <- output1$covParm
covParm

### plot random effects for schools

plotObject(analyticObject=output1)

###############################################
## MLM analysis of cluster randomised trials ##	 
## with bootstrap confidence intervals       ##
###############################################

output2 <- crtFREQ(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",nBoot=1000,data=crtData)


### Effect size

ES2 <- output2$ES
ES2

### plot bootstrapped values

plotObject(analyticObject=output2, group=1)

#######################################################################
## MLM analysis of cluster randomised trials with permutation p-value##
#######################################################################

output3 <- crtFREQ(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",nPerm=1000,data=crtData)


#### Distribution under the null

perm <- output3$Perm


### plot permutated values

plotObject(analyticObject=output3, group=1)


}


}

