% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eefAnalyticPerm.R
\name{mstFREQ}
\alias{mstFREQ}
\title{Analysis of multisite andomised trials using multilevel model.}
\usage{
mstFREQ(formula, random, intervention, nPerm = NULL, data, nBoot = NULL)
}
\arguments{
\item{formula}{specifies in the model using form posttest ~ pretests+Intervention +....}

\item{random}{a string variable specifying the "clustering"  as contained in the data.}

\item{intervention}{specifies the intervention variable in the formula.}

\item{nPerm}{specifies number of perumation under the null hypothesis.  Default is NULL.}

\item{data}{specifies data frame containing the data to be analysed.}

\item{nBoot}{specifies number of bootstrap samples for non-parametric bootstrapping. Default is NULL.}
}
\value{
S3 object; a list consisting of
\itemize{
\item \code{Beta}. The parameter estimates from linear model with the associatie confidence intervals.
\item \code{ES}. A matric of  Hedges effect size based on within variance and total variance. It is base is based on standard error (SE) if \code{nBoot=NULL}. Otherwise, it is based on non-parametric bootstrap confidence intervals.
\item \code{covParm}. Vector of variance decomposition into between-variance (Schools) and within-variance (Pupils).
\item \code{SchEffects}. Estimated random effects for each school.
\item \code{Perm}. A matrix containing the distribution of effect size under the null hypothesis. It is only produced if \code{nPerm} is specified. The first columns contains effect size based on within variance and the second column contains effect size based on total variance.
\item \code{Bootstrap}. A matrix containing the bootstrap effect sizes. It is only prduced is \code{nBoot} is specified. The first columns contains effect size based on within variance and the second column contains effect size based on total variance.
}
}
\description{
\code{crtFREQ} performs analysis of cluster randomised trials.
}
\examples{
data(catcht)

#######################################################################
### Analysis of multisite randomised trials using Hedges Effect size ##
#######################################################################

output1 <- mstFREQ(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",data=catcht)

ES1 <- output1$ES$Intervention1
ES1

ICC1 <- sum(output1$covParm[-3])/sum(output1$covParm)
ICC1

##############################################################
## Analysis of multisite randomised trials with permutation ##
## under the null hypothesis                                ##
##############################################################


output2 <- mstFREQ(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",nPerm=1000,data=catcht)

ES2 <- output2$ES$Intervention1
ES2





ICC2 <- sum(output2$covParm[-3])/sum(output2$covParm)
ICC2

##############################################
### Analysis of multisite randomised trials ##
##############################################


output3 <- mstFREQ(Posttest~ Intervention+Prettest,random="School",
		intervention="Intervention",nBoot=1000,data=catcht)

ES3 <- output3$ES$Intervention1
ES3


ICC3 <- sum(output3$covParm[-3])/sum(output3$covParm)
ICC3

}

