% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefAnalytics_03_2017.r
\name{srtFREQ}
\alias{srtFREQ}
\title{Analysis of Simple Randomised Education Trial using Linear Regression Model.}
\usage{
srtFREQ(formula, intervention, nBoot = NULL, nPerm = NULL, data)
}
\arguments{
\item{formula}{specifies the model to be analysed. It is of the form y~x1+x2+....
Where y is the outcome variable and X?s are the predictors.}

\item{intervention}{specification of the intervention variable as appeared in formula.
This must be put in quotes.  For example ?intervention? or ?treatment? or ?group?.}

\item{nBoot}{number of bootstraps required to generate bootstrap confidence interval. Default is NULL.}

\item{nPerm}{number of permutations required to generate permutation p-value.  Default is NULL.}

\item{data}{data frame containing the data to be analysed.}
}
\value{
S3 object; a list consisting of
\itemize{
\item \code{Beta}. Estimates and confidence intervals for the predictors specified in the model.
\item \code{ES}. Hedges' g for the intervention effect. If nBoot is not specified, the confidence intervals are 95% CIs based on standard errors. If nBoot is specified, they are non-parametric bootstrapped confidence intervals.
\item \code{sigma2}. Residual variance.
\item \code{Perm}. A vector containing the distribution of effect sizes under the null hypothesis. It is produced only if \code{nPerm} is specified.
\item \code{Bootstrap}. A vector containing bootstrapped effect sizes. It is prduced only if \code{nBoot} is specified.
}
}
\description{
\code{srtFREQ} perfoms analysis of educational trials under the assumption of independent errors among pupils. 
This can also be used with schools as fixed effects.
}
\examples{
if(interactive()){

data(mstData)

###################################################################
## Analysis of simple randomised trials using Hedges Effect Size ##
###################################################################

output1 <- srtFREQ(Posttest~ Intervention+Prettest,
		intervention="Intervention",data=mstData )
ES1 <- output1$ES

###################################################################
## Analysis of simple randomised trials using Hedges Effect Size ## 
## with Permutation p-value                                      ##
###################################################################

output2 <- srtFREQ(Posttest~ Intervention+Prettest,
		intervention="Intervention",nPerm=1000,data=mstData )
ES2 <- output2$ES




#### plot permutated values

plotObject(analyticObject=output2, group=1)



###################################################################
## Analysis of simple randomised trials using Hedges Effect Size ##
## with non-parametric bootstrap CI                              ##
###################################################################

output3 <- srtFREQ(Posttest~ Intervention+Prettest,
		intervention="Intervention",nBoot=1000,data=mstData)
ES3 <- output3$ES

### plot bootstrapped values

plotObject(analyticObject=output3, group=1)

####################################################################
## Analysis of simple randomised trials using Hedges Effect Size. ##
## Schools as fixed effects                                       ##
####################################################################

output4 <- srtFREQ(Posttest~ Intervention+Prettest+as.factor(School),
		intervention="Intervention",data=mstData )
ES4 <- output4$ES

###################################################################
## Analysis of simple randomised trials using Hedges Effect Size ##
## with Permutation p-value. Schools as fixed effects            ##
###################################################################

output5 <- srtFREQ(Posttest~ Intervention+Prettest,
		intervention="Intervention",nPerm=1000,data=mstData )

ES5 <- output5$ES
ES5

#### plot permutated values

plotObject(analyticObject=output5, group=1)

####################################################################
## Analysis of simple randomised trials using Hedges Effect Size  ##
## with non-parametric bootstrap CI. Schools as fixed effects     ##
####################################################################

output6 <- srtFREQ(Posttest~ Intervention+Prettest,
		intervention="Intervention",nBoot=1000,data=mstData)
ES6 <- output6$ES
ES6 

### plot bootstrapped values

plotObject(analyticObject=output6, group=1)

}
}

