% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eefAnalyticPerm_modified_05_02_2016.R
\name{srtFREQ}
\alias{srtFREQ}
\title{Analysis of Simple Randomised Trial (SRT).}
\usage{
srtFREQ(formula, intervention, nBoot = NULL, nPerm = NULL, data)
}
\arguments{
\item{formula}{specifies the model to be analysed. It is of the form y~x1+x2+..., 
where y is the outcome variable and X's are the predictors.}

\item{intervention}{the name of the intervention variable as appeared in formula.
This must be put in quotes.  For example "intervention" or "treatment" or "group".}

\item{nBoot}{number of bootstrap required to generate bootstrap confidence interval. Default is NULL.}

\item{nPerm}{number of permutations required to generate permutation p-value. Default  is NULL.}

\item{data}{data frame containing the data to be analysed.}
}
\value{
S3 \code{mcpi} object; a list consisting of
\itemize{
\item \code{Beta}. Estimates and confidence intervals for the predictors specified in the model. 
It will be the slope for a continuous predictor and the mean difference for a dummy variable or a categorical predictor.
\item \code{ES}. Hedges' effect size for the intervention effect. 
If \code{nBoot} is not specified, the confidence intervals are classical 95% CIs based on standard error. 
If \code{nBoot} is specified, they are non-parametric bootstrapped confidence intervals.
\item \code{sigma2}. Residual variance. Its square root will generate a pooled standard deviation. 
\item \code{Perm}. A vector containing the distribution of effect size under the null hypothesis. It is produced only if \code{nPerm} is specified.
}
}
\description{
\code{srtFREQ} perfoms analysis of education trials under the assumption of independent errors between pupils. 
This can also be used with schools as fixed effects.
}
\examples{
data(catcht)

###################################################################
## Analysis of simple randomised trials using Hedges Effect Size ##
###################################################################

output1 <- srtFREQ(Posttest~ Intervention+Prettest,
		intervention="Intervention",data=catcht )
ES1 <- output1$ES

###################################################################
## Analysis of simple randomised trials using Hedges Effect Size ## 
## with Permutation p-value                                      ##
###################################################################

output2 <- srtFREQ(Posttest~ Intervention+Prettest,
		intervention="Intervention",nPerm=1000,data=catcht )
ES2 <- output2$ES




#### Distribution under the null

perm <- output2$Perm


#### Permutation P-value using total variance


obsg <- output2$ES[1]


permPvalue <- ifelse(mean(perm$permES> obsg[1])==0,"<0.001",
		mean(perm$permES > obsg[1]) )
permPvalue 

## Distribution of ES under H0
hist(perm$permES, breaks=40, col="white", border="blueviolet", 
		xlab="Distribution Under Null Hypothesis", 
		main=paste("P(X|NULL)= ",permPvalue,sep=""),
		xlim=range(c(perm[,1],obsg),na.rm=TRUE));
abline(v=obsg[1],lwd=2,col=4)



###################################################################
## Analysis of simple randomised trials using Hedges Effect Size ##
## with non-parametric bootstrap CI                              ##
###################################################################

output3 <- srtFREQ(Posttest~ Intervention+Prettest,
		intervention="Intervention",nBoot=1000,data=catcht)
ES3 <- output3$ES


####################################################################
## Analysis of simple randomised trials using Hedges Effect Size. ##
## Schools as fixed effects                                       ##
####################################################################

output4 <- srtFREQ(Posttest~ Intervention+Prettest+as.factor(School),
		intervention="Intervention",data=catcht )
ES4 <- output4$ES

###################################################################
## Analysis of simple randomised trials using Hedges Effect Size ##
## with Permutation p-value. Schools as fixed effects            ##
###################################################################

output5 <- srtFREQ(Posttest~ Intervention+Prettest,
		intervention="Intervention",nPerm=1000,data=catcht )

ES5 <- output5$ES
ES5

####################################################################
## Analysis of simple randomised trials using Hedges Effect Size  ##
## with non-parametric bootstrap CI. Schools as fixed effects     ##
####################################################################

output6 <- srtFREQ(Posttest~ Intervention+Prettest,
		intervention="Intervention",nBoot=1000,data=catcht)
ES6 <- output6$ES
ES6 
}

